refactor(modal, button, selelect, overlay, card): fix minor issues/typos
