+++ packages/select/components/Button.tsx
@@ -1,7 +1,7 @@
 // @FIXME this needs to be imoprted from 'bl-button' package, not relatively
 
-import { customElement } from '../../_helpers';
-import Button from '../../button/Button';
+import { customElement } from '@blaze-elements/common';
+import { Button as OriginalButton} from '@blaze-elements/button';
 
 @customElement('bl-select-button')
-export class SelectButton extends Button {}
+export default class Button extends OriginalButton {}
